Refactor message configuration: Introduce _RpcMessageConfig class and update popup display level handling, resolivng comments
